requirements.txt
# Dependencias de auditoría
# chia-internal-custody>=0.3.0  # CLI de custodia, instalar manualmente si se necesita

# Procesamiento de datos
pandas>=1.5,<2.0
numpy>=1.23,<2.0
python-dotenv>=0.19.0,<2.0

# Base de datos
psycopg2-binary>=2.9,<3.0
sqlalchemy>=1.4,<2.0

# Utilidades
requests>=2.26.0,<3.0
python-dateutil>=2.8.2,<3.0
tqdm>=4.62.0,<5.0

# Testing
# pytest>=6.2.5,<7.0
# pytest-cov>=2.12.1,<3.0